insights/requirements.txt
streamlit==1.28.1
pandas==2.1.3
plotly==5.17.0
azure-identity==1.14.0
azure-monitor-query==1.2.1
python-dotenv==1.0.0
requests==2.31.0
